allow "+" before a timezone in request log tests
GitHub Actions runs in UTC where the timezone starts with a "+" instead
of our local dev machines (in PT or ET) which start with a "-" in
numeric form. So, we expand the regex to allow for those in
WebSocketRequestLogTest.

Along the way, we change from calling `matches` directly on the strings
to calling it on the `assertThat` to get better test output.